+++ CHANGELOG.md
@@ -15,7 +15,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Expected resolver arguments in `ResolveInfo::enhanceBuilder()`
 - Pass the path array to `CacheKeyAndTags::key()` https://github.com/nuwave/lighthouse/pull/2176
 - Require implementations of `BatchedEntityResolver` to maintain the keys given in `array $representations` https://github.com/nuwave/lighthouse/pull/2286
-- Require filter directives such as `@eq` in `@delete`, `@forceDelete` and `@restore`
+- Require filter directives such as `@whereKey` in `@delete`, `@forceDelete` and `@restore`
 
 ### Fixed